Allow compilation from inside Julia
